+++ apps/docs-v2/data/components/error/ErrorDefault.tsx
@@ -1,5 +1,5 @@
 import {Button, Error, Icon} from '@comfortdelgro/react-compass'
-import { faBug } from '@fortawesome/free-solid-svg-icons'
+import {faBug} from '@fortawesome/free-solid-svg-icons'
 
 function BasicExample() {
   return (