+++ apps/docs-v2/pages/components/quantity-toggle.tsx
@@ -1,0 +1,6 @@
+import MarkdownDocs from 'components/common/MarkdownDocs'
+import * as pageProps from 'docs/data/components/quantity-toggle/quantity-toggle.md?@comfortdelgro/compass-design'
+
+export default function Page() {
+  return <MarkdownDocs {...pageProps} />
+}