+++ packages/react-compass/src/confirm-pickup/confirm-pickup.stories.tsx
@@ -1,48 +1,86 @@
 import React from 'react'
-import Card from '../card'
+import Button from '../button'
+import Drawer from '../drawer'
 import ConfirmPUPoint from './confirm-pickup'
 
-export const Default = () => {
-  const initialPickupPoint = {
-    name: 'Comfort Building',
-    address: '383 Sin Ming Drive',
-    city: 'Singapore',
-    street: '',
-    postalCode: '575717',
-    specificPUPoint: 'Comfort Transporation Lobby',
-  }
-
-  return (
-    <Card css={{backgroundColor: 'lightgray', width: '400px', padding: '10px'}}>
-      <ConfirmPUPoint
-        pickupPoint={initialPickupPoint}
-        onConfirmPickup={(pickupPoint) => {
-          console.log(pickupPoint)
-        }}
-        onChangePUPoint={() => {}}
-      />
-    </Card>
-  )
+const samplePickupPoint = {
+  name: 'Comfort Building',
+  address: '383 Sin Ming Drive',
+  city: 'Singapore',
+  street: '',
+  postalCode: '575717',
+  specificPUPoint: 'Comfort Transporation Lobby',
 }
-export const WithoutSpecificPUPoint = () => {
-  const initialPickupPoint = {
-    name: 'Comfort Building',
-    address: '383 Sin Ming Drive',
-    city: 'Singapore',
-    street: '',
-    postalCode: '575717',
-    specificPUPoint: '',
-  }
-
+export const Variants = () => {
+  const [drawerOpen, setDrawerOpen] = React.useState<boolean>(false)
+  const [drawerNoPUOpen, setDrawerNoPUOpen] = React.useState<boolean>(false)
   return (
-    <Card css={{backgroundColor: 'lightgray', width: '400px', padding: '10px'}}>
-      <ConfirmPUPoint
-        pickupPoint={initialPickupPoint}
-        onConfirmPickup={(pickupPoint) => {
-          console.log(pickupPoint)
+    <>
+      <h4>Confirm PU Point with specific location</h4>
+      <Button
+        onPress={() => {
+          setDrawerOpen(true)
         }}
-        placeholder='E.g. Lobby, side entrance, etc'
-      />
-    </Card>
+      >
+        Open Drawer
+      </Button>
+      <h4>Confirm PU Point without specific location</h4>
+      <Button
+        onPress={() => {
+          setDrawerNoPUOpen(true)
+        }}
+      >
+        Open Drawer
+      </Button>
+      <Drawer
+        variant='h5'
+        disableResize
+        open={drawerOpen}
+        onClose={() => setDrawerOpen(false)}
+        css={{
+          height: '35dvh',
+        }}
+      >
+        <ConfirmPUPoint
+          puPointTitle={samplePickupPoint.name}
+          puPointDescription={`${samplePickupPoint.address}, ${samplePickupPoint.city} ${samplePickupPoint.postalCode}`}
+          specificPUPoint={samplePickupPoint.specificPUPoint}
+          onConfirmPickup={(pickupPoint) => {
+            alert('confirm the pickup point')
+          }}
+          onChangePUPoint={() => {
+            alert('change the pickup point')
+          }}
+          css={{
+            padding: 0,
+            borderRadius: 0,
+          }}
+        />
+      </Drawer>
+
+      <Drawer
+        variant='h5'
+        disableResize
+        open={drawerNoPUOpen}
+        onClose={() => setDrawerNoPUOpen(false)}
+        css={{
+          height: '40dvh',
+        }}
+      >
+        <ConfirmPUPoint
+          puPointTitle={samplePickupPoint.name}
+          puPointDescription={`${samplePickupPoint.address}, ${samplePickupPoint.city} ${samplePickupPoint.postalCode}`}
+          specificPUPoint={''}
+          onConfirmPickup={(pickupPoint) => {
+            alert(`confirm the pickup point ${pickupPoint}`)
+          }}
+          onChangePUPoint={() => {
+            alert('change the pickup point')
+          }}
+          placeholder='E.g. Lobby, side entrance, etc'
+          css={{padding: 0, borderRadius: 0}}
+        />
+      </Drawer>
+    </>
   )
 }