+++ packages/react-compass/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## 2.9.0
 
+## 2.15.0
+
+### Minor Changes
+
+- df06bff9: Fix minor bugs
+
+## 2.14.0
+
+### Minor Changes
+
+- eb5e467b: Fix minor bugs and add new h5 components
+
 ## 2.13.0
 
 ### Minor Changes